Added DS elastic agent
