+++ deepspeed/elasticity/__init__.py
@@ -1,1 +1,2 @@
 from .elasticity import compute_elastic_config, elasticity_enabled, ensure_immutable_elastic_config
+from .elastic_agent import DSElasticAgent
