+++ deepspeed/launcher/runner.py
@@ -70,10 +70,22 @@ def parse_args(args=None):
                         ''')
 
     parser.add_argument("--num_nodes",
-                        type=str,
-                        default="-1",
+                        type=int,
+                        default=-1,
                         help="Total number of worker nodes to run on, this will use "
                         "the top N hosts from the given hostfile.")
+
+    parser.add_argument("--min_num_nodes",
+                        type=int,
+                        default=-1,
+                        help="Minimum number of nodes to run elastic training on. "
+                        "Default is 1 when elastic training is enabled")
+
+    parser.add_argument("--max_num_nodes",
+                        type=int,
+                        default=-1,
+                        help="Maximum number of nodes to run elastic training on. "
+                        "Default is num_nodes when elastic training is enabled")
 
     parser.add_argument("--num_gpus",
                         type=int,
@@ -335,8 +347,6 @@ def parse_num_nodes(str_num_nodes: str, elastic_training: bool):
 def main(args=None):
     args = parse_args(args)
 
-    num_nodes, args.max_num_nodes = parse_num_nodes(args.num_nodes, args.elastic_training)
-
     if args.elastic_training:
         assert args.master_addr != "", "Master Addr is required when elastic training is enabled" 
 
@@ -347,7 +357,7 @@ def main(args=None):
     if not resource_pool and len(cuda_visible_devices):
         detected_str = f"Detected CUDA_VISIBLE_DEVICES={cuda_visible_devices}"
         if len(args.include) or len(
-                args.exclude) or num_nodes > 1 or args.num_gpus > 0:
+                args.exclude) or args.num_nodes > 1 or args.num_gpus > 0:
             print(
                 f"{detected_str} but ignoring it because one or several of --include/--exclude/--num_gpus/--num_nodes cl args were used. If you want to use CUDA_VISIBLE_DEVICES don't pass any of these arguments to deepspeed."
             )
@@ -356,7 +366,7 @@ def main(args=None):
             print(f"{detected_str}: setting --include={args.include}")
         del os.environ["CUDA_VISIBLE_DEVICES"]
 
-    if num_nodes >= 0 or args.num_gpus >= 0:
+    if args.num_nodes >= 0 or args.num_gpus >= 0:
         if args.include != "" or args.exclude != "":
             raise ValueError("Cannot specify num_nodes/gpus with include/exclude")
 
@@ -370,7 +380,7 @@ def main(args=None):
         args.master_addr = "127.0.0.1"
         multi_node_exec = False
 
-    if not multi_node_exec and num_nodes > 1:
+    if not multi_node_exec and args.num_nodes > 1:
         raise ValueError("Num nodes is >1 but no extra nodes available via hostfile")
 
     active_resources = parse_inclusion_exclusion(resource_pool,
@@ -404,10 +414,10 @@ def main(args=None):
         run_autotuning(args, active_resources)
         return
 
-    if num_nodes > 0:
+    if args.num_nodes > 0:
         updated_active_resources = collections.OrderedDict()
         for count, hostname in enumerate(active_resources.keys()):
-            if num_nodes == count:
+            if args.num_nodes == count:
                 break
             updated_active_resources[hostname] = active_resources[hostname]
         active_resources = updated_active_resources
@@ -444,6 +454,7 @@ def main(args=None):
         if args.elastic_training:
             deepspeed_launch.append("--enable_elastic_training")
             deepspeed_launch.append(f"--max_nodes={args.max_num_nodes}")
+            deepspeed_launch.append(f"--min_nodes={args.min_num_nodes}")
         cmd = deepspeed_launch + [args.user_script] + args.user_args
     else:
         args.launcher = args.launcher.lower()
